GCloud/tacotron_hifigan/requirements.txt
--find-links https://download.pytorch.org/whl/cpu
torch
torchaudio


Flask==3.0.3
functions_framework==3.8.0
numpy==2.0.0
soundfile==0.12.1
sox==1.5.0
google-cloud-storage==2.5.0